backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
httpx==0.25.2
requests==2.31.0 